refactor(a2a): restructure A2A modules into focused submodules
Split the monolithic A2A handler/filters into four focused submodules
(discovery, jsonrpc, streaming, rest) matching the OpenAI adapter pattern.
Move FileAgentCardRegistry from bitrouter-a2a to bitrouter runtime where
other concrete implementations live. Group A2A runtime files into a2a/
subdirectory for consistency with the API-side restructuring.
