+++ bitrouter-api/src/router/a2a/jsonrpc/dispatch.rs
@@ -1,14 +1,8 @@
-//! JSON-RPC 2.0 dispatch handler for A2A server methods.
-//!
-//! Handles all A2A v1.0 methods: `SendMessage`, `GetTask`, `CancelTask`,
-//! `ListTasks`, `GetExtendedAgentCard`, `SendStreamingMessage`,
-//! `SubscribeToTask`, and push notification config CRUD.
+//! JSON-RPC 2.0 method dispatch for A2A server methods.
 
 use std::sync::Arc;
-use std::sync::atomic::{AtomicU64, Ordering};
 
-use bitrouter_a2a::error::A2aError;
-use bitrouter_a2a::jsonrpc::{JsonRpcError, JsonRpcRequest, JsonRpcResponse};
+use bitrouter_a2a::jsonrpc::{JsonRpcRequest, JsonRpcResponse};
 use bitrouter_a2a::registry::AgentCardRegistry;
 use bitrouter_a2a::request::{
     CancelTaskRequest, DeleteTaskPushNotificationConfigRequest,
@@ -20,6 +14,11 @@ use bitrouter_a2a::server::{
 };
 use bitrouter_a2a::stream::StreamResponse;
 use bitrouter_a2a::task::{GetTaskRequest, ListTasksRequest, ListTasksResponse, Task};
+
+use super::convert::{
+    GetExtendedByFirst, WithId, deserialize_params, error_response, execution_error_response,
+    generate_id, success_response,
+};
 
 /// Dispatch a JSON-RPC request to the appropriate A2A method handler.
 pub async fn handle_jsonrpc<E, S, R, P>(
@@ -315,286 +314,10 @@ where
         Ok(()) => success_response(&request.id, &serde_json::json!({"success": true})),
         Err(e) => {
             let code = match &e {
-                A2aError::PushNotificationNotFound { .. } => -32001,
+                bitrouter_a2a::error::A2aError::PushNotificationNotFound { .. } => -32001,
                 _ => -32603,
             };
             error_response(&request.id, code, e.to_string())
-        }
-    }
-}
-
-// ── REST handler wrappers ───────────────────────────────────────
-
-/// Handle REST `POST /message:send` — wraps `SendMessage`.
-pub async fn handle_rest_send_message<E, S>(
-    body: SendMessageRequest,
-    executor: Arc<E>,
-    task_store: Arc<S>,
-) -> Box<dyn warp::Reply>
-where
-    E: AgentExecutor,
-    S: TaskStore,
-{
-    let task_id = body
-        .message
-        .task_id
-        .clone()
-        .unwrap_or_else(|| generate_id("task"));
-    let context_id = body
-        .message
-        .context_id
-        .clone()
-        .unwrap_or_else(|| generate_id("ctx"));
-
-    let ctx = ExecutorContext {
-        message: body.message,
-        task_id,
-        context_id,
-        configuration: body.configuration,
-    };
-
-    match executor.execute(&ctx).await {
-        Ok(ExecuteResult::Task(task)) => {
-            let _ = task_store.create(&task);
-            Box::new(warp::reply::with_status(
-                warp::reply::json(&task),
-                warp::http::StatusCode::OK,
-            ))
-        }
-        Ok(ExecuteResult::Message(msg)) => Box::new(warp::reply::with_status(
-            warp::reply::json(&msg),
-            warp::http::StatusCode::OK,
-        )),
-        Err(e) => Box::new(warp::reply::with_status(
-            warp::reply::json(&serde_json::json!({"error": e.to_string()})),
-            warp::http::StatusCode::INTERNAL_SERVER_ERROR,
-        )),
-    }
-}
-
-/// Handle REST `GET /tasks/{id}`.
-pub fn handle_rest_get_task<S>(task_id: String, task_store: Arc<S>) -> Box<dyn warp::Reply>
-where
-    S: TaskStore,
-{
-    match task_store.get(&task_id) {
-        Ok(Some(stored)) => Box::new(warp::reply::json(&stored.task)),
-        Ok(None) => Box::new(warp::reply::with_status(
-            warp::reply::json(&serde_json::json!({"error": format!("task not found: {task_id}")})),
-            warp::http::StatusCode::NOT_FOUND,
-        )),
-        Err(e) => Box::new(warp::reply::with_status(
-            warp::reply::json(&serde_json::json!({"error": e.to_string()})),
-            warp::http::StatusCode::INTERNAL_SERVER_ERROR,
-        )),
-    }
-}
-
-/// Handle REST `POST /tasks/{id}:cancel`.
-pub async fn handle_rest_cancel_task<E, S>(
-    task_id: String,
-    executor: Arc<E>,
-    task_store: Arc<S>,
-) -> Box<dyn warp::Reply>
-where
-    E: AgentExecutor,
-    S: TaskStore,
-{
-    match task_store.get(&task_id) {
-        Ok(Some(_)) => {}
-        Ok(None) => {
-            return Box::new(warp::reply::with_status(
-                warp::reply::json(
-                    &serde_json::json!({"error": format!("task not found: {task_id}")}),
-                ),
-                warp::http::StatusCode::NOT_FOUND,
-            ));
-        }
-        Err(e) => {
-            return Box::new(warp::reply::with_status(
-                warp::reply::json(&serde_json::json!({"error": e.to_string()})),
-                warp::http::StatusCode::INTERNAL_SERVER_ERROR,
-            ));
-        }
-    }
-
-    match executor.cancel(&task_id).await {
-        Ok(task) => {
-            let _ = task_store.create(&task);
-            Box::new(warp::reply::json(&task))
-        }
-        Err(e) => Box::new(warp::reply::with_status(
-            warp::reply::json(&serde_json::json!({"error": e.to_string()})),
-            warp::http::StatusCode::INTERNAL_SERVER_ERROR,
-        )),
-    }
-}
-
-/// Handle REST `POST /tasks/{id}/push-notification-configs`.
-pub fn handle_rest_create_push_config<P>(
-    _task_id: String,
-    config: TaskPushNotificationConfig,
-    push_store: Arc<P>,
-) -> Box<dyn warp::Reply>
-where
-    P: PushNotificationStore,
-{
-    match push_store.create_config(&config) {
-        Ok(stored) => Box::new(warp::reply::with_status(
-            warp::reply::json(&stored),
-            warp::http::StatusCode::CREATED,
-        )),
-        Err(e) => Box::new(warp::reply::with_status(
-            warp::reply::json(&serde_json::json!({"error": e.to_string()})),
-            warp::http::StatusCode::INTERNAL_SERVER_ERROR,
-        )),
-    }
-}
-
-/// Handle REST `GET /tasks/{id}/push-notification-configs/{config_id}`.
-pub fn handle_rest_get_push_config<P>(
-    task_id: String,
-    config_id: String,
-    push_store: Arc<P>,
-) -> Box<dyn warp::Reply>
-where
-    P: PushNotificationStore,
-{
-    match push_store.get_config(&task_id, &config_id) {
-        Ok(Some(config)) => Box::new(warp::reply::json(&config)),
-        Ok(None) => Box::new(warp::reply::with_status(
-            warp::reply::json(&serde_json::json!({
-                "error": format!("push config not found: task={task_id} id={config_id}")
-            })),
-            warp::http::StatusCode::NOT_FOUND,
-        )),
-        Err(e) => Box::new(warp::reply::with_status(
-            warp::reply::json(&serde_json::json!({"error": e.to_string()})),
-            warp::http::StatusCode::INTERNAL_SERVER_ERROR,
-        )),
-    }
-}
-
-/// Handle REST `GET /tasks/{id}/push-notification-configs`.
-pub fn handle_rest_list_push_configs<P>(task_id: String, push_store: Arc<P>) -> Box<dyn warp::Reply>
-where
-    P: PushNotificationStore,
-{
-    match push_store.list_configs(&task_id) {
-        Ok(configs) => {
-            let response = ListTaskPushNotificationConfigsResponse { configs };
-            Box::new(warp::reply::json(&response))
-        }
-        Err(e) => Box::new(warp::reply::with_status(
-            warp::reply::json(&serde_json::json!({"error": e.to_string()})),
-            warp::http::StatusCode::INTERNAL_SERVER_ERROR,
-        )),
-    }
-}
-
-/// Handle REST `DELETE /tasks/{id}/push-notification-configs/{config_id}`.
-pub fn handle_rest_delete_push_config<P>(
-    task_id: String,
-    config_id: String,
-    push_store: Arc<P>,
-) -> Box<dyn warp::Reply>
-where
-    P: PushNotificationStore,
-{
-    match push_store.delete_config(&task_id, &config_id) {
-        Ok(()) => Box::new(warp::reply::with_status(
-            warp::reply::json(&serde_json::json!({"success": true})),
-            warp::http::StatusCode::OK,
-        )),
-        Err(A2aError::PushNotificationNotFound { .. }) => Box::new(warp::reply::with_status(
-            warp::reply::json(&serde_json::json!({
-                "error": format!("push config not found: task={task_id} id={config_id}")
-            })),
-            warp::http::StatusCode::NOT_FOUND,
-        )),
-        Err(e) => Box::new(warp::reply::with_status(
-            warp::reply::json(&serde_json::json!({"error": e.to_string()})),
-            warp::http::StatusCode::INTERNAL_SERVER_ERROR,
-        )),
-    }
-}
-
-// ── Helpers ─────────────────────────────────────────────────────
-
-fn deserialize_params<T: serde::de::DeserializeOwned>(
-    params: &serde_json::Value,
-) -> Result<T, Box<JsonRpcResponse>> {
-    serde_json::from_value::<T>(params.clone())
-        .map_err(|e| Box::new(error_response("", -32602, format!("invalid params: {e}"))))
-}
-
-fn success_response<T: serde::Serialize>(id: &str, result: &T) -> JsonRpcResponse {
-    let value = serde_json::to_value(result).unwrap_or(serde_json::Value::Null);
-    JsonRpcResponse {
-        jsonrpc: "2.0".to_string(),
-        id: id.to_string(),
-        result: Some(value),
-        error: None,
-    }
-}
-
-fn error_response(id: &str, code: i64, message: String) -> JsonRpcResponse {
-    JsonRpcResponse {
-        jsonrpc: "2.0".to_string(),
-        id: id.to_string(),
-        result: None,
-        error: Some(JsonRpcError {
-            code,
-            message,
-            data: None,
-        }),
-    }
-}
-
-fn execution_error_response(id: &str, err: &A2aError) -> JsonRpcResponse {
-    let code = match err {
-        A2aError::TaskNotFound { .. } => -32001,
-        A2aError::Execution(_) => -32000,
-        _ => -32603,
-    };
-    error_response(id, code, err.to_string())
-}
-
-fn generate_id(prefix: &str) -> String {
-    static COUNTER: AtomicU64 = AtomicU64::new(1);
-    format!("{prefix}-{}", COUNTER.fetch_add(1, Ordering::Relaxed))
-}
-
-/// Helper to set the id on an error response built without one.
-trait WithId {
-    fn with_id(self, id: &str) -> Self;
-}
-
-impl WithId for JsonRpcResponse {
-    fn with_id(mut self, id: &str) -> Self {
-        self.id = id.to_string();
-        self
-    }
-}
-
-/// Helper trait for `GetExtendedAgentCard` — finds first registered agent.
-trait GetExtendedByFirst {
-    fn get_extended_by_first(
-        &self,
-        registry: &dyn AgentCardRegistry,
-    ) -> Result<Option<bitrouter_a2a::registry::AgentRegistration>, A2aError>;
-}
-
-impl<R: AgentCardRegistry> GetExtendedByFirst for R {
-    fn get_extended_by_first(
-        &self,
-        _registry: &dyn AgentCardRegistry,
-    ) -> Result<Option<bitrouter_a2a::registry::AgentRegistration>, A2aError> {
-        let list = self.list()?;
-        if let Some(first) = list.into_iter().next() {
-            self.get_extended(&first.card.name)
-        } else {
-            Ok(None)
         }
     }
 }
