+++ bitrouter-api/src/router/a2a/mod.rs
@@ -1,2 +1,7 @@
-pub mod filters;
-pub mod handler;
+pub mod discovery;
+pub mod jsonrpc;
+pub mod rest;
+pub mod streaming;
+
+#[cfg(test)]
+pub(crate) mod test_helpers;
