+++ bitrouter/Cargo.toml
@@ -88,5 +88,8 @@ solana-pubkey = { version = "4.1", optional = true }
 solana-transaction = { version = "3.1", optional = true }
 spl-associated-token-account = { version = "8.0", optional = true, default-features = false }
 
+[dev-dependencies]
+tempfile = { version = "3" }
+
 [target.'cfg(unix)'.dependencies]
 libc = "0.2"
